+++ src/Services/ResourceService.php
@@ -183,6 +183,15 @@ class ResourceService implements ResourceServiceInterface
             case $resourceType === IdStrings::CANCEL:
                 $paymentId  = IdService::getResourceIdFromUrl($url, IdStrings::PAYMENT);
                 $chargeId   = IdService::getResourceIdOrNullFromUrl($url, IdStrings::CHARGE);
+                if (IdService::isPaymentCancellation($url)) {
+                    $isRefund = preg_match('/charge/', $url) === 1;
+                    if ($isRefund) {
+                        $resource = $unzer->fetchPaymentRefund($paymentId, $resourceId);
+                        break;
+                    }
+                    $resource = $unzer->fetchPaymentReversal($paymentId, $resourceId);
+                    break;
+                }
                 if ($chargeId !== null) {
                     $resource = $unzer->fetchRefundById($paymentId, $chargeId, $resourceId);
                     break;
@@ -781,6 +790,36 @@ class ResourceService implements ResourceServiceInterface
         $cancel = $this->fetchResource($charge->getCancellation($cancellationId, true));
         return $cancel;
     }
+
+    /**
+     * {@inheritDoc}
+     */
+    public function fetchPaymentRefund($payment, $cancellationId): Cancellation
+    {
+        $charge = new Charge();
+        $charge->setParentResource($payment);
+        $cancel = (new Cancellation())
+            ->setId($cancellationId)
+            ->setPayment($payment)
+            ->setParentResource($charge);
+        return $this->fetchResource($cancel);
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    public function fetchPaymentReversal($payment, $cancellationId): Cancellation
+    {
+        $authorization = new Authorization();
+        $authorization->setParentResource($payment);
+        $cancel = (new Cancellation())
+            ->setId($cancellationId)
+            ->setPayment($payment)
+            ->setParentResource($authorization);
+        return $this->fetchResource($cancel);
+    }
+
+
 
     //</editor-fold>