+++ test/unit/Services/PaymentServiceTest.php
@@ -182,21 +182,55 @@ class PaymentServiceTest extends BasePaymentTest
     }
 
     /**
+     * @deprecated
      * Verify chargeAuthorization calls fetchPayment if the payment object is passed as id string.
      *
      * @test
      */
-    public function chargeAuthorizationShouldCallFetchPaymentIfThePaymentIsPassedAsIdString(): void
+    public function chargeAuthorizationShouldCallChargePaymentMethod(): void
     {
         /** @var ResourceServiceInterface|MockObject $resourceSrvMock */
         $resourceSrvMock = $this->getMockBuilder(ResourceService::class)->setMethods(['fetchPayment'])->disableOriginalConstructor()->getMock();
-        $resourceSrvMock->expects($this->once())->method('fetchPayment')->willReturn(new Payment());
         /** @var PaymentService|MockObject $paymentSrvMock */
-        $paymentSrvMock = $this->getMockBuilder(PaymentService::class)->setMethods(['chargePayment', 'getResourceService'])->disableOriginalConstructor()->getMock();
+        $paymentSrvMock = $this->getMockBuilder(PaymentService::class)->setMethods(['chargePayment'])->disableOriginalConstructor()->getMock();
         $paymentSrvMock->expects($this->once())->method('chargePayment')->withAnyParameters();
-        $paymentSrvMock->expects(self::once())->method('getResourceService')->willReturn($resourceSrvMock);
 
         $paymentSrvMock->chargeAuthorization('myPaymentId');
+    }
+
+    /**
+     * @deprecated
+     * Verify chargePayment calls fetchPayment if the payment object is passed as id string.
+     *
+     * @test
+     */
+    public function chargePaymentShouldCallerformChargeOnPaymentMethod(): void
+    {
+        /** @var ResourceServiceInterface|MockObject $resourceSrvMock */
+        $resourceSrvMock = $this->getMockBuilder(ResourceService::class)->setMethods(['fetchPayment'])->disableOriginalConstructor()->getMock();
+        /** @var PaymentService|MockObject $paymentSrvMock */
+        $paymentSrvMock = $this->getMockBuilder(PaymentService::class)->setMethods(['performChargeOnPayment'])->disableOriginalConstructor()->getMock();
+        $paymentSrvMock->expects($this->once())->method('performChargeOnPayment')->withAnyParameters();
+
+        $paymentSrvMock->chargePayment('myPaymentId');
+    }
+
+    /**
+     * Verify performChargeOnPayment calls fetchPayment if the payment object is passed as id string.
+     *
+     * @test
+     */
+    public function performChargeOnPaymentShouldCallFetchPaymentIfThePaymentIsPassedAsIdString(): void
+    {
+        /** @var ResourceServiceInterface|MockObject $resourceSrvMock */
+        $resourceSrvMock = $this->getMockBuilder(ResourceService::class)->setMethods(['fetchPayment', 'createResource'])->disableOriginalConstructor()->getMock();
+        $resourceSrvMock->expects($this->once())->method('fetchPayment')->willReturn(new Payment());
+        $resourceSrvMock->expects($this->once())->method('createResource')->willReturn(new Charge());
+        /** @var PaymentService|MockObject $paymentSrvMock */
+        $paymentSrvMock = $this->getMockBuilder(PaymentService::class)->setMethods(['getResourceService'])->disableOriginalConstructor()->getMock();
+        $paymentSrvMock->expects(self::exactly(2))->method('getResourceService')->willReturn($resourceSrvMock);
+
+        $paymentSrvMock->performChargeOnPayment('myPaymentId', new Charge());
     }
 
     /**
@@ -218,7 +252,6 @@ class PaymentServiceTest extends BasePaymentTest
                 $newPayment = $charge->getPayment();
                 return $charge instanceof Charge &&
                     $charge->getAmount() === 1.234 &&
-                    $charge->getCurrency() === 'myTestCurrency' &&
                     $charge->getOrderId() === null &&
                     $charge->getInvoiceId() === null &&
                     $newPayment instanceof Payment &&
@@ -227,7 +260,7 @@ class PaymentServiceTest extends BasePaymentTest
             }));
 
         $paymentSrv     = $unzer->setResourceService($resourceSrvMock)->getPaymentService();
-        $returnedCharge = $paymentSrv->chargePayment($payment, 1.234, 'myTestCurrency');
+        $returnedCharge = $paymentSrv->chargePayment($payment, 1.234);
         $this->assertEquals([$returnedCharge], $payment->getCharges());
     }
 
@@ -250,7 +283,6 @@ class PaymentServiceTest extends BasePaymentTest
                 $newPayment = $charge->getPayment();
                 return $charge instanceof Charge &&
                     $charge->getAmount() === 1.234 &&
-                    $charge->getCurrency() === 'myTestCurrency' &&
                     $charge->getOrderId() === 'orderId' &&
                     $charge->getInvoiceId() === 'invoiceId' &&
                     $newPayment instanceof Payment &&
@@ -259,7 +291,7 @@ class PaymentServiceTest extends BasePaymentTest
             }));
 
         $paymentSrv     = $unzer->setResourceService($resourceSrvMock)->getPaymentService();
-        $returnedCharge = $paymentSrv->chargePayment($payment, 1.234, 'myTestCurrency', 'orderId', 'invoiceId');
+        $returnedCharge = $paymentSrv->chargePayment($payment, 1.234, 'orderId', 'invoiceId');
         $this->assertEquals([$returnedCharge], $payment->getCharges());
     }