+++ CHANGELOG.md
@@ -3,14 +3,24 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [1.1.6.0](https://github.com/unzerdev/php-sdk/compare/1.1.5.0..1.1.6.0)
+## [1.2.0.0](https://github.com/unzerdev/php-sdk/compare/1.1.5.0..1.2.0.0)
 ### Added
 * Add payment type Paylater Invoice.
 * Add properties `companyType` and `owner` to `CompanyInfo` class.
 * Add shipping type to `Address` class.
 * Allow setting the clientIp manually.
+* Allow setting riskData for authorize request.
+* Allow setting shipping for charge request.
+* Add new methods for `authorize` and `charge` transactions that use prepared objects only.
+* Add new methods to cancel payments done via paylater-invoice type:
+  * `Unzer::cancelAuthorizedPayment()`
+  * `Unzer::cancelChargedPayment()`
+* Add new methods to fetch cancellations of payment done via paylater-invoice type:
+  * `Unzer::fetchPaymentReversal()`
+  * `Unzer::fetchPaymentRefund()`
 
 ### Changed
+* Remove redundant `currency` parameter from `Unzer::chargePayment()` method.
 * Add "geoLocation" property to all payment type classes.
 * Several minor improvements.